+++ src/app/(providers)/@modal/(.)mungstagram/[id]/page.module.scss
@@ -1,0 +1,12 @@
+.modal {
+  padding: 15px;
+  width: 400px;
+  height: 600px;
+  position: absolute;
+  top: 15%;
+  left: 41%;
+  justify-content: center;
+  background-color: white;
+  border-radius: 10px;
+  z-index: 1000;
+}
