+++ src/app/(providers)/@modal/(.)mungstagram/create/page.module.scss
@@ -1,7 +1,3 @@
-// .modal {
-//   z-index: 10;
-// }
-
 .modal {
   padding: 15px;
   width: 400px;
