+++ src/app/(providers)/@modal/(.)mungstagram/create/page.tsx
@@ -2,6 +2,7 @@
 
 import ReactModal from 'react-modal';
 import { useRouter } from 'next/navigation';
+import styles from './page.module.scss';
 
 const MungModal = () => {
   const router = useRouter();
@@ -11,8 +12,14 @@ const MungModal = () => {
   };
 
   return (
-    <ReactModal isOpen={true} onRequestClose={closeModal} ariaHideApp={false} contentLabel="Modal">
-      <div>게시글 상세 모달</div>
+    <ReactModal
+      className={styles.modal}
+      isOpen={true}
+      onRequestClose={closeModal}
+      ariaHideApp={false}
+      contentLabel="Modal"
+    >
+      <div>게시글 작성 모달</div>
     </ReactModal>
   );
 };
